der: extract `length` module
Small cleanup and consistency improvement.

It may be possible to eliminate this module in the future, or it might
make sense to make it the one place for all length-related functionality
which could be useful.

For now, it gets this code out of the way to permit more refactoring.
